Make Palace's own linalg/fem/models code compile under MFEM_USE_SINGLE
The actual source-code work identified as the real blocker in
docs/src/developer/apple-metal-occa-progress.md section 9: Palace's
own code hardcodes double in numerous places that alias, override, or
bind by reference to mfem::real_t-typed data, which only ever worked
because real_t == double in every build to date. Under
MFEM_USE_SINGLE (real_t == float), each of these is a hard compile
error, not a slow path. Fixed each one found while getting a
PALACE_PRECISION=single build to compile end-to-end:

- palace/linalg/hypre.cpp: three sites doing
  const_cast<double *>(mfem::real_t*) to feed hypre_VectorData /
  hypre_CSRMatrixData (raw HYPRE_Real* fields) -- the original bug
  that started this investigation. Fixed to HYPRE_Real, the type the
  struct fields actually are.
- palace::Operator::AddMult/AddMultTranspose overrides across
  linalg/hypre.hpp, linalg/rap.hpp, linalg/operator.hpp
  (BaseDiagonalOperator, BaseProductOperator, BaseMultigridOperator,
  SumOperatorTraits<Operator>), and fem/libceed/operator.hpp: all
  override mfem::Operator::AddMult(..., const real_t a), but declared
  the scalar parameter as double, silently breaking `override`
  (caught by the compiler, not a runtime bug -- these simply failed
  to build).
- palace/fem/coefficient.hpp: seven mfem::Coefficient::Eval overrides
  (and their out-of-line InterfaceDielectricCoefficient<> template
  specializations) declared `double Eval(...)  override` against a
  base that declares `virtual real_t Eval(...)`.
- Local `double foo_data[N]; mfem::Vector foo(foo_data, N);`-style
  stack arrays across fem/coefficient.hpp and
  models/waveportoperator.cpp (aliases real_t-backed Vector storage).
- palace/linalg/vector.hpp: StaticVector<N>'s backing buffer.
- palace/models/materialoperator.hpp: mat_c0_min/mat_c0_max, an
  mfem::Array<double> whose Min()/Max() calls only link against
  MFEM's explicitly-instantiated Array<real_t> in a single-precision
  MFEM build (a link-time-only symptom, not a compile error).
- palace/models/timeoperator.cpp: ImplicitSolve/SUNImplicitSetup/
  SUNImplicitSolve overrides of mfem::TimeDependentOperator virtuals
  (same double-vs-real_t override mismatch as above); TimeOperator::
  Step bridges its own double&-typed public interface to MFEM's
  ODESolver::Step(real_t&, real_t&) via local real_t temporaries
  rather than changing Step's own signature (Step's double t/dt
  matches the driver's own double-typed waveform evaluation, which is
  deliberately left double -- see the models/*port*.cpp entries
  below).
- palace/linalg/densematrix.cpp, palace/models/floquetportoperator.cpp,
  palace/linalg/solver.cpp, palace/fem/libceed/operator.cpp,
  palace/linalg/vector.cpp: assorted real_t/double arithmetic-type
  mismatches (std::max, mfem::Array2D<real_t> for
  HypreParMatrixFromBlocks, a templated lambda instead of a
  double-only one in ComplexVector::Set, etc.).
- palace/models/lumpedportoperator.cpp, palace/models/waveportoperator.cpp:
  std::complex<double> * mfem::real_t has no operator* overload
  (complex<T>'s scalar operators require an exact T match); added
  explicit static_cast<double>(...) at each site -- these compute a
  genuinely double-precision complex mode-overlap integral by design
  (Palace's ComplexVector/EigenvalueSolver interfaces are
  deliberately double throughout, independent of real_t), so casting
  up is correct, not a workaround.

None of these change behavior for double-precision builds:
mfem::real_t == double there, so every fix above is a no-op rename,
not a logic change. Verified via -fsyntax-only against the real
(double) MFEM/HYPRE headers for every file touched.

+++ palace/models/lumpedportoperator.cpp
@@ -259,7 +259,8 @@ std::complex<double> LumpedPortData::GetPower(GridFunction &E, GridFunction &B) 
     pr.UseDevice(false);
     pr.Assemble();
     pr.UseDevice(true);
-    dot = -(pr * E.Real()) + (has_imag ? -1i * (pr * E.Imag()) : 0.0);
+    dot = -static_cast<double>(pr * E.Real()) +
+          (has_imag ? -1i * static_cast<double>(pr * E.Imag()) : 0.0);
   }
   if (has_imag)
   {
@@ -269,7 +270,8 @@ std::complex<double> LumpedPortData::GetPower(GridFunction &E, GridFunction &B) 
     pi.UseDevice(false);
     pi.Assemble();
     pi.UseDevice(true);
-    dot += -(pi * E.Imag()) + 1i * (pi * E.Real());
+    dot += -static_cast<double>(pi * E.Imag()) +
+          1i * static_cast<double>(pi * E.Real());
     Mpi::GlobalSum(1, &dot, E.ParFESpace()->GetComm());
     return dot;
   }
